Issue-230 - Split ECS and ALB into new stack
