+++ src/cdk-stack/src/main.ts
@@ -1,5 +1,6 @@
 import * as cdk from 'aws-cdk-lib';
 import { EcrStack } from './ecr-stack';
+import { EcsClusterStack } from './ecs-cluster-stack';
 import { LambdaDispatchStack } from './lambda-dispatch-stack';
 import { VpcStack } from './vpc-stack';
 
@@ -15,25 +16,29 @@ new EcrStack(app, 'ecr-stack', {
   env: devEnv,
   stackName: 'lambda-dispatch-ecr',
 });
-
-// Only create VPC stack and use its VPC for the main app stack
 const vpcStack = new VpcStack(app, 'vpc-stack', {
   env: devEnv,
-  stackName: 'vpc-with-nat-instances',
+  stackName: 'lambda-dispatch-vpc',
 });
+new EcsClusterStack(app, 'ecs-stack', {
+  env: devEnv,
+  stackName: 'lambda-dispatch-ecs',
+  vpc: vpcStack.vpc,
+});
+
 const lambdaDispatchStack = new LambdaDispatchStack(app, 'lambda-dispatch', {
   env: devEnv,
   stackName: 'lambda-dispatch-app-stack',
   vpc: vpcStack.vpc,
-  loadBalancerArn: cdk.Fn.importValue('vpc-with-nat-instances-LoadBalancerArn'),
-  httpsListenerArn: cdk.Fn.importValue('vpc-with-nat-instances-HttpsListenerArn'),
+  loadBalancerArn: cdk.Fn.importValue('lambda-dispatch-ecs-LoadBalancerArn'),
+  httpsListenerArn: cdk.Fn.importValue('lambda-dispatch-ecs-HttpsListenerArn'),
   loadBalancerSecurityGroupId: cdk.Fn.importValue(
-    'vpc-with-nat-instances-LoadBalancerSecurityGroupId',
+    'lambda-dispatch-ecs-LoadBalancerSecurityGroupId',
   ),
-  loadBalancerDnsName: cdk.Fn.importValue('vpc-with-nat-instances-ALBDnsName'),
-  loadBalancerHostedZoneId: cdk.Fn.importValue('vpc-with-nat-instances-ALBCanonicalHostedZoneId'),
-  ecsClusterArn: cdk.Fn.importValue('vpc-with-nat-instances-ClusterArn'),
-  ecsClusterName: cdk.Fn.importValue('vpc-with-nat-instances-ClusterName'),
+  loadBalancerDnsName: cdk.Fn.importValue('lambda-dispatch-ecs-ALBDnsName'),
+  loadBalancerHostedZoneId: cdk.Fn.importValue('lambda-dispatch-ecs-ALBCanonicalHostedZoneId'),
+  ecsClusterArn: cdk.Fn.importValue('lambda-dispatch-ecs-ClusterArn'),
+  ecsClusterName: cdk.Fn.importValue('lambda-dispatch-ecs-ClusterName'),
 });
 cdk.Tags.of(lambdaDispatchStack).add('Name', 'lambda-dispatch');
 
@@ -41,15 +46,15 @@ const lambdaDispatchStackPr = new LambdaDispatchStack(app, 'lambda-dispatch-pr',
   env: devEnv,
   stackName: `lambda-dispatch-app-stack-pr-${process.env.PR_NUMBER}`,
   vpc: vpcStack.vpc,
-  loadBalancerArn: cdk.Fn.importValue('vpc-with-nat-instances-LoadBalancerArn'),
-  httpsListenerArn: cdk.Fn.importValue('vpc-with-nat-instances-HttpsListenerArn'),
+  loadBalancerArn: cdk.Fn.importValue('lambda-dispatch-ecs-LoadBalancerArn'),
+  httpsListenerArn: cdk.Fn.importValue('lambda-dispatch-ecs-HttpsListenerArn'),
   loadBalancerSecurityGroupId: cdk.Fn.importValue(
-    'vpc-with-nat-instances-LoadBalancerSecurityGroupId',
+    'lambda-dispatch-ecs-LoadBalancerSecurityGroupId',
   ),
-  loadBalancerDnsName: cdk.Fn.importValue('vpc-with-nat-instances-ALBDnsName'),
-  loadBalancerHostedZoneId: cdk.Fn.importValue('vpc-with-nat-instances-ALBCanonicalHostedZoneId'),
-  ecsClusterArn: cdk.Fn.importValue('vpc-with-nat-instances-ClusterArn'),
-  ecsClusterName: cdk.Fn.importValue('vpc-with-nat-instances-ClusterName'),
+  loadBalancerDnsName: cdk.Fn.importValue('lambda-dispatch-ecs-ALBDnsName'),
+  loadBalancerHostedZoneId: cdk.Fn.importValue('lambda-dispatch-ecs-ALBCanonicalHostedZoneId'),
+  ecsClusterArn: cdk.Fn.importValue('lambda-dispatch-ecs-ClusterArn'),
+  ecsClusterName: cdk.Fn.importValue('lambda-dispatch-ecs-ClusterName'),
   removalPolicy: cdk.RemovalPolicy.DESTROY,
 });
 cdk.Tags.of(lambdaDispatchStackPr).add('Name', `lambda-dispatch-pr-${process.env.PR_NUMBER}`);
